category-service 에러 처리 수정

+++ src/services/category-service.js
@@ -16,15 +16,24 @@ class CategoryService {
   }
 
   async getCategories() {
-    return await this.categoryModel.findAll();
+    const categories = await this.categoryModel.findAll();
+    if (!categories) {
+      throw new Error("카테고리 내역이 없습니다. 다시 한 번 확인해 주세요.");
+    }
+    return categories;
+
   }
 
   async getCategoryById(categoryId) {
-    return await this.categoryModel.findById(categoryId);
+    const category = await this.categoryModel.findById(categoryId);
+    if (!category) {
+      throw new Error("카테고리 내역이 없습니다. 다시 한 번 확인해 주세요.");
+    }
+    return category;
   }
 
   async setCategory(categoryId, toUpdate) {
-    //우선 해당 id의 상품이 db에 있는지 확인
+    //우선 해당 id의 카테고리가 db에 있는지 확인
     let category = await this.categoryModel.findById(categoryId);
 
     //db에서 찾지 못한 경우, 에러 메시지 반환
@@ -39,6 +48,9 @@ class CategoryService {
   }
 
   async deleteCategory(categoryIdArray) {
+    if (categoryIdArray.length < 1) {
+      throw new Error("삭제할 카테고리가 없습니다. 다시 한 번 확인해 주세요.");
+    }
     let category = await categoryIdArray.map((categoryId) =>
       this.categoryModel.deleteById({ categoryId })
     );
